レポートのUIを改善
